# openapi.yaml
openapi: 3.0.0
info:
  title: Enterprise AI Incident Commander API
  description: API used by IBM watsonx Orchestrate to manage enterprise incidents
  version: "1.0.0"

servers:
  - url: http://localhost:5050

paths:
  /incident:
    get:
      summary: Fetch a new enterprise incident
      operationId: getIncident
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                type: object
                properties:
                  incident_id:
                    type: string
                  system:
                    type: string
                  description:
                    type: string
                  business_impact:
                    type: string
                  signals:
                    type: array
                    items:
                      type: string
                  timestamp:
                    type: string

  /assign-owner:
    post:
      summary: Assign an owner based on severity
      operationId: assignOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                incident_id:
                  type: string
                severity:
                  type: string
      responses:
        "200":
          description: Assigned owner
          content:
            application/json:
              schema:
                type: object
                properties:
                  incident_id:
                    type: string
                  severity:
                    type: string
                  assigned_owner:
                    type: string

  /notify:
    post:
      summary: Send incident update to stakeholders
      operationId: notifyStakeholders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Notification result